refactor(kwin-effect): extract paint pipeline and shader transitions
Final Phase 1 step. Splits the paint+shader chunk of plasmazoneseffect.cpp
into two TUs sharing a private header for hot-path constexpr arrays
(plasmazoneseffect/shader_internal.h):

- paint_pipeline.cpp: prePaintScreen, postPaintScreen, prePaintWindow,
  paintWindow.
- shader_transitions.cpp: evictLruTextureIfOverBound,
  warmUserTextureAsync, beginShaderTransition, endShaderTransition,
  tryBeginShaderForEvent, loadShaderProfileFromDbus,
  loadShaderRegistryFromDbus. Moves the anon-namespace shader helpers
  (injectKwinDefineAfterVersion, loadUserTextureImage, wrapStringToEnum)
  into this TU's anon namespace. Calls
  PhosphorProtocol::ClientHelpers::loadSettingAsync directly rather
  than the daemon_bringup-local member-template wrapper, so the
  template body doesn't need to be visible here.

The main plasmazoneseffect.cpp shrinks from 1844 lines to 88 — it now
holds only the Q_LOGGING_CATEGORY definition, the five effect-interface
methods, and the KWIN_EFFECT_FACTORY_SUPPORTED macro + .moc include.

Phase 1 final layout:
  plasmazoneseffect.cpp                         88
  plasmazoneseffect/shader_transitions.cpp    1226 (over 800 — gated on
                                                    beginShaderTransition's
                                                    function-level decomposition;
                                                    parking-lot follow-up)
  plasmazoneseffect/daemon_bringup.cpp         732
  plasmazoneseffect/daemon_apply.cpp           694
  plasmazoneseffect/lifecycle.cpp              688
  plasmazoneseffect/drag_snap.cpp              643
  plasmazoneseffect/window_lifecycle.cpp       625
  plasmazoneseffect/paint_pipeline.cpp         506
  plasmazoneseffect/screens.cpp                425
  plasmazoneseffect/window_filtering.cpp       254
  plasmazoneseffect/borders.cpp                182
  plasmazoneseffect/mouse_drag.cpp             138
  plasmazoneseffect/window_identity.cpp        136
  plasmazoneseffect/shader_internal.h           62

Down from a single 6075-line plasmazoneseffect.cpp.
